backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
langchain==0.2.0
langchain-upstage==0.1.0
pillow==10.3.0
httpx==0.27.0
python-dotenv==1.0.1
